move some lib files to lib/cocoapods-packager folder
