+++ lib/cocoapods-packager/builder.rb
