+++ lib/cocoapods-packager/framework.rb
