+++ lib/cocoapods-packager/mangle.rb
