+++ lib/cocoapods-packager/pod_utils.rb
