+++ lib/cocoapods-packager/spec_builder.rb
