+++ lib/cocoapods-packager/symbols.rb
