+++ lib/cocoapods_plugin.rb
@@ -1,7 +1,7 @@
 require 'pod/command/package'
-require 'builder'
-require 'framework'
-require 'mangle'
-require 'pod_utils'
-require 'spec_builder'
-require 'symbols'
+require 'cocoapods-packager/builder'
+require 'cocoapods-packager/framework'
+require 'cocoapods-packager/mangle'
+require 'cocoapods-packager/pod_utils'
+require 'cocoapods-packager/spec_builder'
+require 'cocoapods-packager/symbols'
